moved createObject{Input,Output}Stream() into ManagerBase to avoid casting
